rustc_mir: move all analyses to an analysis module.
