+++ src/librustc_mir/analysis/mod.rs
@@ -8,13 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-pub mod borrowck_errors;
-pub mod elaborate_drops;
-pub mod patch;
-
-mod graphviz;
-pub(crate) mod pretty;
-
-pub use self::pretty::{dump_enabled, dump_mir, write_mir_pretty, PassWhere};
-pub use self::graphviz::{write_mir_graphviz};
-pub use self::graphviz::write_node_label as write_graphviz_node_label;
+pub mod alignment;
+pub mod dataflow;
+pub mod def_use;
+pub mod liveness;
